+++ CHANGELOG.md
@@ -6,8 +6,9 @@ Changes since 1.4.1:
 
 ### Overall
 
-- Update **.NET Core** version to **3.1.101**;
+- Update **.NET Core** version to **3.1.102**;
 - Fixed **Linux** build and apps;
+- Create **Docker image** for **GTiff2Tiles.Console**;
 
 ### Core